mtd: ndmpart.c: use octal permissions instead of symbolic
 WARNING: Symbolic permissions 'S_IRUGO | S_IWUSR' are not preferred.
 Consider using octal permissions '0644'.
 #1195: FILE: drivers/mtd/ndmpart.c:1195:
 +		entry = proc_create(DI_BOOT_ACTIVE, S_IRUGO | S_IWUSR, proc_dir,

 WARNING: Symbolic permissions 'S_IRUGO | S_IWUSR' are not preferred.
 Consider using octal permissions '0644'.
 #1199: FILE: drivers/mtd/ndmpart.c:1199:
 +		entry = proc_create(DI_BOOT_BACKUP, S_IRUGO | S_IWUSR, proc_dir,

 WARNING: Symbolic permissions 'S_IRUGO | S_IWUSR' are not preferred.
 Consider using octal permissions '0644'.
 #1203: FILE: drivers/mtd/ndmpart.c:1203:
 +		entry = proc_create("boot_current", S_IRUGO | S_IWUSR, proc_dir,

 WARNING: Symbolic permissions 'S_IRUGO | S_IWUSR' are not preferred.
 Consider using octal permissions '0644'.
 #1207: FILE: drivers/mtd/ndmpart.c:1207:
 +		entry = proc_create(DI_BOOT_FAILS, S_IRUGO | S_IWUSR, proc_dir,

 WARNING: Symbolic permissions 'S_IWUSR' are not preferred.
 Consider using octal permissions '0200'.
 #1211: FILE: drivers/mtd/ndmpart.c:1211:
 +		entry = proc_create("commit", S_IWUSR, proc_dir, &fops_commit);

"... It's *much* easier to parse and understand the octal
numbers, while the symbolic macro names are just random line noise and
hard as hell to understand. You really have to think about it.

So we should rather go the other way: convert existing bad symbolic
permission bit macro use to just use the octal numbers.

The symbolic names are good for the *other* bits (ie sticky bit, and
the inode mode _type_ numbers etc), but for the permission bits, the
symbolic names are just insane crap. Nobody sane should ever use them.
Not in the kernel, not in user space.

           Linus"